добавит разделение на модули и подлкючение их в точке входа
