+++ js/const.js
@@ -3,7 +3,7 @@ const MIN_QUNTITY_LIKES = 15;
 const MAX_QUNTITY_LIKES = 200;
 const QUNTITY_COMMETS = 30;
 const QUNTITY_AVATAR_COMMETS = 6;
-const MAX_ID_COMMENTS = 300;
+const MAX_ID_COMMENTS = QUNTITY_COMMETS * QUNTITY_POST;
 
 
 export {QUNTITY_POST, MIN_QUNTITY_LIKES, MAX_QUNTITY_LIKES, QUNTITY_COMMETS, QUNTITY_AVATAR_COMMETS, MAX_ID_COMMENTS};
